rename t1 to struct and mandatory brain volume value in registration functions
