+++ sammba/registration/struct.py
